requirements.txt
Flask
google-cloud-storage
orjson
PyYAML
requests